Initial Upload
Some of this code hasn't been touched in a few years. I have a few simple goals for it, like fixing the archive div spacing and adding some buttons in the footer of posts to allow easy navigation between blog entries.

I also would like to add a day/night mode and a way to hide shitposts from my employers!